requirements.txt
asgiref==3.5.2
django-filebrowser-no-grappelli==4.0.1
django-filter==21.1
django-model-utils==4.2.0
django-widget-tweaks==1.4.12
Django==4.0.4
flake8==4.0.1
mccabe==0.6.1
mysqlclient==2.1.0
Pillow==9.1.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
